requirements.txt
pip>=23.1
python-dotenv>=1.0.0
requests>=2.28.2
setuptools>=58.1.0
flask